Corrigida a remoção de comentários

+++ app/controllers/comments_controller.rb
@@ -1,20 +1,18 @@
 class CommentsController < ApplicationController
-  
   def create
     file = UserFile.find(params[:comment][:file])
-    file.comments << Comment.create(:rate => params[:comment][:rate], 
-                                    :message => params[:comment][:message], 
-                                    :owner => current_user,
-                                    :file => file)
+    file.comments << Comment.create(:rate => params[:comment][:rate],
+                                     :message => params[:comment][:message],
+                                     :owner => current_user,
+                                     :file => file)
     redirect_to :back
   end
-  
+
   def destroy
-    comment = Comment.find(params[:id])
-    if comment.owner == current_user && comment.owner != nil
-      comment.destroy
-    end
+    file = UserFile.find(params[:file_id])
+    comment = file.comments.find(params[:id])
+    comment.destroy unless comment.owner != current_user || comment.owner == nil
     redirect_to :back
   end
-  
+
 end
